rename registration forms
